pyproject.toml
[project]
name = "attendance-system-v2"
version = "2.0.0"
description = "Distributed Facial-Recognition Attendance System V2"
requires-python = ">=3.12"
dependencies = [
    # --- Core Framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.10.0",
    "python-multipart>=0.0.20",

    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",

    # --- Auth & Security ---
    "PyJWT>=2.10.0",
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.3.0",

    # --- Cache & Task Queue ---
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",

    # --- AI / CV Pipeline ---
    "numpy>=2.2.0",
    "opencv-python>=4.11.0",
    "pillow>=11.3.0",
    "sahi>=0.11.34",
    "ultralytics>=8.3.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.22.0",
    "gfpgan>=1.3.8",

    # --- Scheduling ---
    "apscheduler>=3.10.4",

    # --- Observability ---
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
